requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.1
sqlparse==0.4.2
yapf==0.32.0
